fix param type

+++ core/client/src/main/java/net/bfsr/client/particle/SpawnAccumulator.java
@@ -19,7 +19,7 @@ public class SpawnAccumulator {
         accumulatedTime += (emitTime - lastEmitTime) * 0.000000001;
     }
 
-    public void consume(float spawnTime) {
+    public void consume(double spawnTime) {
         accumulatedTime -= spawnTime;
     }
 }
